requirements.txt
langgraph==0.2.45
langchain-core>=0.3.17,<0.4.0
httpx==0.27.2
python-dotenv==1.0.1
watchdog==5.0.3
pydantic>=2.0.0

# Optional: Only needed if you want to integrate with OpenAI LLMs
# langchain-openai==0.2.8
